According to talk with Tomeu:
I change terminology to make bulmafact more generic.
English: from "company" to "establishment"
Spanish: from "empresa" to "entidad"
Catalan: from "empresa" to "entitat"

So that entering basic data for the establishment looks OK no matter if it is 
a firm, an association or any other for profit or non for profit organitzation.